Add Dolma Counter script
This PR adds a new dolma processor that can be used to count the number
of (whitespace delimited) tokens in a data source.

You can point it at a directory and it will find all the `.jsonl.gz`
files in it or it's subdirectories.

It can be used from the root dir via `python -m licensed_pile.count` or
from anywhere with `count-tokens-dolma`.