Tool for comparing dolma data after preprocessing. (#86)
Data is read (and cached), then it can be navigated with next/prev to
move long as if scrolling though the file on disk. On can also jump to a
specific page via id or title (both which support type-ahead search). It
also allows for jumping to a random page for spot checking. Once an
example is selected, the two versions are shown to the user side-by-side