Feat/tokenizer (#104)
* tokenizer

* update to handle raw jsonl.gz files

* first try at tokenizer preprocessor

* ups

* fix streaming datasets

* Update to tokenizer for usage with lingua.

The hf_to_tiktoken utilities can be used to dump our hf BPE tokenizer
into a format that can be loaded/used with TikToken.

The hf_lingua file is an implementation of the lingua tokenizer using
an HF tokenizer as the backend. To use in lingua we will need to update
the `build_tokenizer` function to include the ability to import and
instantiate our tokenizer.

I also updated the training scripts to make it easier to include
pre-tokenization regex like the ones used in llama or gpt2. I also
added a version of the llama regex that splits individual digits instead
of groups of 3.

* fix how they convert bpe data to vocab

* Added script that verifies converted tokenizers.

* added tests on more sources

* ups

* ups

* add cli version of the conversion script